Updates after renaming PushNotificationRegistrationStatus class
